requirements.txt
streamlit
langchain-groq
langchain-community
python-dotenv
google-search-results
